+++ include/Camera.h
@@ -60,7 +60,7 @@ class Camera {
     static glm::mat4 projection;
     static glm::mat4 view;
     static float rotationDeltaX, rotationDeltaY;
-    static bool updateViewFrustum, dirty, showOverlay;
+    static bool updateViewFrustum, dirty, showOverlay, movingFaster;
     static int room;
 };